seperate the fetch process
